pyproject.toml
[tool.poetry]
name = "LocalLLMBot"
version = "0.1.0"
description = ""
authors = ["dvnguyen02 <duynguyen290502@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ollama = "^0.3.3"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
